Preserve team fees natively from next-week generation through publishing
